Revert "changes to make @BHoMBot happy"
This reverts commit 702721155b65399cd36d03a7335830facc916a1e.
